Removed unnecessary lines of code in change log and controller also renamed files
