+++ Microservices/ServiceRegistryMicroservice/src/main/java/cs4337/group_8/ServiceRegistry/ServiceRegistryApplication.java
@@ -1,4 +1,4 @@
-package cs4337.group_8.TemplateProject;
+package cs4337.group_8.ServiceRegistry;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
