+++ Microservices/ServiceRegistryMicroservice/src/main/resources/db/changelog-master.xml
@@ -4,10 +4,6 @@
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.4.xsd">
 
-<!--    We will have a parent directory -->
-    <include file="com/example/news/news.changelog.sql"/>
-    <include file="com/example/directory/directory.changelog.sql"/>
-
     <changeSet author="Killian" id="0">
         <!-- Creating comments table -->
         <createTable tableName="comments">
