+++ Microservices/ServiceRegistryMicroservice/src/test/java/cs4337/group_8/ServiceRegistry/ServiceRegistryApplicationTests.java
@@ -1,10 +1,10 @@
-package cs4337.group_8.TemplateProject;
+package cs4337.group_8.ServiceRegistry;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class TemplateProjectApplicationTests {
+class ServiceRegistryApplicationTests {
 
 	@Test
 	void contextLoads() {
